@@ -12,7 +12,7 @@ authors = [
 ]
 readme = "README.md"
 license = {text = "GPL-3.0"}
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
     "attrs",
     "aiohttp",
@@ -20,7 +20,6 @@ dependencies = [
     "async_timeout",
     "crccheck",
     "cryptography",
-    'importlib_resources; python_version<"3.9"',
     'async-timeout; python_version<"3.11"',
     "voluptuous",
     "jsonschema",
@@ -58,7 +57,7 @@ asyncio_mode = "auto"
 
 [tool.ruff]
 required-version = ">=0.5.0"
-target-version = "py38"
+target-version = "py39"
 line-length = 88
 
 exclude = [
