@@ -3,9 +3,9 @@
 from __future__ import annotations
 
 import enum
-from typing import TYPE_CHECKING, Any, Dict, Union
+from typing import TYPE_CHECKING, Any, Union
 
-ConfigType = Dict[str, Any]
+ConfigType = dict[str, Any]
 
 # pylint: disable=invalid-name
 ClusterType = "Cluster"
@@ -40,7 +40,7 @@ if TYPE_CHECKING:
     ClusterType = zigpy.zcl.Cluster
     ControllerApplicationType = zigpy.application.ControllerApplication
     CustomClusterType = zigpy.quirks.CustomCluster
-    CustomDeviceType = zigpy.quirks.CustomDevice
+    CustomDeviceType = zigpy.quirks.BaseCustomDevice
     CustomEndpointType = zigpy.quirks.CustomEndpoint
     DeviceType = zigpy.device.Device
     EndpointType = zigpy.endpoint.Endpoint
