@@ -1,7 +1,9 @@
 from __future__ import annotations
 
 import asyncio
+from typing import Any
 from unittest import mock
+from unittest.mock import AsyncMock, MagicMock, patch, sentinel
 
 import pytest
 
@@ -10,13 +12,12 @@ import zigpy.device
 import zigpy.endpoint
 import zigpy.types as t
 from zigpy.zcl import foundation
-
-from .async_mock import AsyncMock, MagicMock, int_sentinel, patch, sentinel
+from zigpy.zcl.clusters.general import Ota
 
 DEFAULT_TSN = 123
 
 
-@pytest.fixture()
+@pytest.fixture
 def endpoint():
     ep = zigpy.endpoint.Endpoint(MagicMock(), 1)
     ep.add_input_cluster(0)
@@ -88,7 +89,7 @@ def test_manufacturer_specific_cluster():
     assert hasattr(c, "cluster_id")
 
 
-@pytest.fixture()
+@pytest.fixture
 def cluster_by_id():
     def _cluster(cluster_id=0):
         epmock = MagicMock()
@@ -101,79 +102,173 @@ def cluster_by_id():
     return _cluster
 
 
-@pytest.fixture()
+@pytest.fixture
 def cluster(cluster_by_id):
     return cluster_by_id(0)
 
 
-@pytest.fixture()
+@pytest.fixture
 def client_cluster():
     epmock = AsyncMock()
     epmock.device.get_sequence = MagicMock(return_value=DEFAULT_TSN)
-    return zcl.Cluster.from_id(epmock, 3)
+    return Ota(epmock)
 
 
 async def test_request_general(cluster):
-    await cluster.request(True, 0, [])
+    await cluster.request(
+        general=True,
+        command_id=foundation.GENERAL_COMMANDS[
+            foundation.GeneralCommand.Read_Attributes
+        ].id,
+        schema=foundation.GENERAL_COMMANDS[
+            foundation.GeneralCommand.Read_Attributes
+        ].schema,
+        attribute_ids=[],
+    )
     assert cluster._endpoint.request.call_count == 1
 
 
 async def test_request_manufacturer(cluster):
-    await cluster.request(True, 0, [t.uint8_t], 1)
+    command = foundation.ZCLCommandDef(
+        name="test_command", id=0x00, schema={"param1": t.uint8_t}
+    ).with_compiled_schema()
+
+    await cluster.request(
+        general=True,
+        command_id=command.id,
+        schema=command.schema,
+        param1=1,
+    )
     assert cluster._endpoint.request.call_count == 1
-    org_size = len(cluster._endpoint.request.call_args[0][2])
-    await cluster.request(True, 0, [t.uint8_t], 1, manufacturer=1)
+
+    org_size = len(cluster._endpoint.request.mock_calls[0].kwargs["data"])
+    await cluster.request(
+        general=True,
+        command_id=command.id,
+        schema=command.schema,
+        param1=1,
+        manufacturer=1,
+    )
     assert cluster._endpoint.request.call_count == 2
-    assert org_size + 2 == len(cluster._endpoint.request.call_args[0][2])
+    assert org_size + 2 == len(cluster._endpoint.request.mock_calls[1].kwargs["data"])
 
 
 async def test_request_optional(cluster):
-    schema = [t.uint8_t, t.uint16_t, t.Optional(t.uint16_t), t.Optional(t.uint8_t)]
+    command = foundation.ZCLCommandDef(
+        name="test_command",
+        id=0x00,
+        schema={
+            "param1": t.uint8_t,
+            "param2": t.uint16_t,
+            "param3?": t.uint16_t,
+            "param4?": t.uint8_t,
+        },
+    ).with_compiled_schema()
+
     cluster.endpoint.request = AsyncMock()
 
     with pytest.raises(ValueError):
-        await cluster.request(True, 0, schema)
+        await cluster.request(
+            general=True,
+            command_id=command.id,
+            schema=command.schema,
+        )
 
     assert cluster._endpoint.request.call_count == 0
     cluster._endpoint.request.reset_mock()
 
     with pytest.raises(ValueError):
-        await cluster.request(True, 0, schema, 1)
+        await cluster.request(
+            general=True,
+            command_id=command.id,
+            schema=command.schema,
+            param1=1,
+        )
 
     assert cluster._endpoint.request.call_count == 0
     cluster._endpoint.request.reset_mock()
 
-    await cluster.request(True, 0, schema, 1, 2)
+    await cluster.request(
+        general=True,
+        command_id=command.id,
+        schema=command.schema,
+        param1=1,
+        param2=2,
+    )
     assert cluster._endpoint.request.call_count == 1
     cluster._endpoint.request.reset_mock()
 
-    await cluster.request(True, 0, schema, 1, 2, 3)
+    await cluster.request(
+        general=True,
+        command_id=command.id,
+        schema=command.schema,
+        param1=1,
+        param2=2,
+        param3=3,
+    )
     assert cluster._endpoint.request.call_count == 1
     cluster._endpoint.request.reset_mock()
 
-    await cluster.request(True, 0, schema, 1, 2, 3, 4)
+    await cluster.request(
+        general=True,
+        command_id=command.id,
+        schema=command.schema,
+        param1=1,
+        param2=2,
+        param3=3,
+        param4=4,
+    )
     assert cluster._endpoint.request.call_count == 1
     cluster._endpoint.request.reset_mock()
 
     with pytest.raises(TypeError):
-        await cluster.request(True, 0, schema, 1, 2, 3, 4, 5)
+        await cluster.request(
+            general=True,
+            command_id=command.id,
+            schema=command.schema,
+            param1=1,
+            param2=2,
+            param3=3,
+            param4=4,
+            param5=5,
+        )
 
     assert cluster._endpoint.request.call_count == 0
     cluster._endpoint.request.reset_mock()
 
 
 async def test_reply_general(cluster):
-    await cluster.reply(False, 0, [])
+    command = foundation.ZCLCommandDef(
+        name="test_command", id=0x00, schema={}
+    ).with_compiled_schema()
+
+    await cluster.reply(general=False, command_id=command.id, schema=command.schema)
     assert cluster._endpoint.reply.call_count == 1
 
 
 async def test_reply_manufacturer(cluster):
-    await cluster.reply(False, 0, [t.uint8_t], 1)
+    command = foundation.ZCLCommandDef(
+        name="test_command",
+        id=0x00,
+        schema={
+            "param1": t.uint8_t,
+        },
+    ).with_compiled_schema()
+
+    await cluster.reply(
+        general=False, command_id=command.id, schema=command.schema, param1=1
+    )
     assert cluster._endpoint.reply.call_count == 1
-    org_size = len(cluster._endpoint.reply.call_args[0][2])
-    await cluster.reply(False, 0, [t.uint8_t], 1, manufacturer=1)
+    org_size = len(cluster._endpoint.reply.mock_calls[0].kwargs["data"])
+    await cluster.reply(
+        general=False,
+        command_id=command.id,
+        schema=command.schema,
+        param1=1,
+        manufacturer=1,
+    )
     assert cluster._endpoint.reply.call_count == 2
-    assert org_size + 2 == len(cluster._endpoint.reply.call_args[0][2])
+    assert org_size + 2 == len(cluster._endpoint.reply.mock_calls[1].kwargs["data"])
 
 
 def test_attribute_report(cluster):
@@ -197,21 +292,10 @@ def test_attribute_report(cluster):
     cluster.handle_message(hdr, cmd)
     assert cluster._attr_cache[attr.attrid] == "manufacturer"
 
-    def mock_type(*args, **kwargs):
-        raise ValueError
-
-    with patch.dict(
-        cluster.attributes,
-        {0xAAAA: foundation.ZCLAttributeDef(id=0xAAAA, name="Name", type=mock_type)},
-    ):
-        attr.attrid = 0xAAAA
-        cluster.handle_message(hdr, cmd)
-        assert cluster._attr_cache[attr.attrid] == "manufacturer"
-
 
 def test_handle_request_unknown(cluster):
     hdr = MagicMock(auto_spec=foundation.ZCLHeader)
-    hdr.command_id = int_sentinel.command_id
+    hdr.command_id = 0x42
     hdr.frame_control.is_general = True
     hdr.frame_control.is_cluster = False
     cluster.listener_event = MagicMock()
@@ -229,10 +313,10 @@ def test_handle_request_unknown(cluster):
 
 def test_handle_cluster_request(cluster):
     hdr = MagicMock(auto_spec=foundation.ZCLHeader)
-    hdr.command_id = int_sentinel.command_id
+    hdr.command_id = 0x42
     hdr.frame_control.is_general = False
     hdr.frame_control.is_cluster = True
-    hdr.command_id.is_general = False
+
     cluster.listener_event = MagicMock()
     cluster._update_attribute = MagicMock()
     cluster.handle_cluster_general_request = MagicMock()
@@ -322,28 +406,6 @@ async def test_read_attributes_default_response(cluster):
     success, failure = await cluster.read_attributes([0, 5, 23], allow_cache=False)
     assert success == {}
     assert failure == {0: 0xC1, 5: 0xC1, 23: 0xC1}
-
-
-async def test_read_attributes_value_normalization_error(cluster):
-    async def mockrequest(
-        foundation, command, schema, args, manufacturer=None, **kwargs
-    ):
-        assert foundation is True
-        assert command == 0
-        rar5 = _mk_rar(5, "Model")
-        return [[rar5]]
-
-    def mock_type(*args, **kwargs):
-        raise ValueError
-
-    cluster.request = mockrequest
-    with patch.dict(
-        cluster.attributes,
-        {5: foundation.ZCLAttributeDef(id=5, name="Name", type=mock_type)},
-    ):
-        success, failure = await cluster.read_attributes(["model"], allow_cache=True)
-    assert failure == {}
-    assert success["model"] == "Model"
 
 
 async def test_item_access_attributes(cluster):
@@ -440,14 +502,14 @@ async def test_write_attributes_raw(cluster):
     ],
 )
 async def test_write_attribute_types(
-    cluster_id, attr, value, serialized, cluster_by_id
+    cluster_id: int, attr: str, value: Any, serialized: bytes, cluster_by_id
 ):
     cluster = cluster_by_id(cluster_id)
     with patch.object(cluster.endpoint, "request", new=AsyncMock()):
         await cluster.write_attributes({attr: value})
         assert cluster._endpoint.reply.call_count == 0
         assert cluster._endpoint.request.call_count == 1
-        assert cluster.endpoint.request.call_args[0][2][3:] == serialized
+        assert cluster.endpoint.request.mock_calls[0].kwargs["data"][3:] == serialized
 
 
 @pytest.mark.parametrize(
@@ -540,57 +602,6 @@ async def test_write_attributes_cache_failure(cluster, attributes, result, faile
                 )
 
 
-async def test_read_attributes_response(cluster):
-    await cluster.read_attributes_rsp({0: 5})
-    assert cluster._endpoint.reply.call_count == 1
-    assert cluster._endpoint.request.call_count == 0
-
-
-async def test_read_attributes_resp_unsupported(cluster):
-    await cluster.read_attributes_rsp({0: 5})
-    assert cluster._endpoint.reply.call_count == 1
-    assert cluster._endpoint.request.call_count == 0
-    orig_len = len(cluster._endpoint.reply.call_args[0][2])
-
-    await cluster.read_attributes_rsp({0: 5, 2: None})
-    assert cluster._endpoint.reply.call_count == 2
-    assert cluster._endpoint.request.call_count == 0
-    assert len(cluster._endpoint.reply.call_args[0][2]) == orig_len + 3
-
-
-async def test_read_attributes_resp_str(cluster):
-    await cluster.read_attributes_rsp({"hw_version": 32})
-    assert cluster._endpoint.reply.call_count == 1
-    assert cluster._endpoint.request.call_count == 0
-
-
-async def test_read_attributes_resp_exc(cluster):
-    with patch.object(foundation.DATA_TYPES, "pytype_to_datatype_id") as mck:
-        mck.side_effect = ValueError
-        await cluster.read_attributes_rsp({"hw_version": 32})
-    assert cluster._endpoint.reply.call_count == 1
-    assert cluster._endpoint.request.call_count == 0
-    assert cluster.endpoint.reply.call_args[0][2][-3:] == b"\x03\x00\x86"
-
-
-@pytest.mark.parametrize(
-    ("cluster_id", "attr", "value", "serialized"),
-    [
-        (0, "zcl_version", 0xAA, b"\x00\x00\x00\x20\xaa"),
-        (0, "model", "model x", b"\x05\x00\x00\x42\x07model x"),
-        (0, "device_enabled", True, b"\x12\x00\x00\x10\x01"),
-        (0, "alarm_mask", 0x55, b"\x13\x00\x00\x18\x55"),
-        (0x0202, "fan_mode", 0xDE, b"\x00\x00\x00\x30\xde"),
-    ],
-)
-async def test_read_attribute_resp(cluster_id, attr, value, serialized, cluster_by_id):
-    cluster = cluster_by_id(cluster_id)
-    await cluster.read_attributes_rsp({attr: value})
-    assert cluster._endpoint.reply.call_count == 1
-    assert cluster._endpoint.request.call_count == 0
-    assert cluster.endpoint.reply.call_args[0][2][3:] == serialized
-
-
 def test_bind(cluster):
     cluster.bind()
 
@@ -667,19 +678,19 @@ async def test_configure_reporting_types(cluster_id, attr, data_type, cluster_by
     await cluster.configure_reporting(attr, 0x1234, 0x2345, 0xAA)
     assert cluster._endpoint.reply.call_count == 0
     assert cluster._endpoint.request.call_count == 1
-    assert cluster.endpoint.request.call_args[0][2][6] == data_type
+    assert cluster.endpoint.request.mock_calls[0].kwargs["data"][6] == data_type
 
 
 async def test_command(cluster):
     await cluster.command(0x00)
     assert cluster._endpoint.request.call_count == 1
-    assert cluster._endpoint.request.call_args[0][1] == DEFAULT_TSN
+    assert cluster._endpoint.request.mock_calls[0].kwargs["sequence"] == DEFAULT_TSN
 
 
 async def test_command_override_tsn(cluster):
     await cluster.command(0x00, tsn=22)
     assert cluster._endpoint.request.call_count == 1
-    assert cluster._endpoint.request.call_args[0][1] == 22
+    assert cluster._endpoint.request.mock_calls[0].kwargs["sequence"] == 22
 
 
 async def test_command_attr(cluster):
@@ -688,7 +699,7 @@ async def test_command_attr(cluster):
 
 
 async def test_client_command_attr(client_cluster):
-    await client_cluster.identify_query_response(timeout=0)
+    await client_cluster.query_specific_file_response(status=foundation.Status.SUCCESS)
     assert client_cluster._endpoint.reply.call_count == 1
 
 
@@ -703,8 +714,13 @@ async def test_invalid_arguments_cluster_command(cluster):
 
 
 async def test_invalid_arguments_cluster_client_command(client_cluster):
-    with pytest.raises(TypeError):
-        await client_cluster.client_command(0, 0, 0)
+    with pytest.raises(ValueError):
+        await client_cluster.client_command(
+            command_id=Ota.ClientCommandDefs.upgrade_end_response.id,
+            manufacturer_code=0,
+            image_type=0,
+            # Missing: file_version, current_time, upgrade_time
+        )
 
 
 def test_name(cluster):
@@ -807,12 +823,20 @@ async def test_write_attributes_undivided(cluster):
 
 
 async def test_configure_reporting_multiple(cluster):
-    await cluster.configure_reporting(3, 5, 15, 20, manufacturer=0x2345)
-    await cluster.configure_reporting_multiple({3: (5, 15, 20)}, manufacturer=0x2345)
+    await cluster.configure_reporting(
+        attribute=3,
+        min_interval=5,
+        max_interval=15,
+        reportable_change=20,
+        manufacturer=0x2345,
+    )
+    await cluster.configure_reporting_multiple(
+        attributes={3: (5, 15, 20)}, manufacturer=0x2345
+    )
     assert cluster.endpoint.request.call_count == 2
     assert (
-        cluster.endpoint.request.call_args_list[0][0][2]
-        == cluster.endpoint.request.call_args_list[1][0][2]
+        cluster.endpoint.request.mock_calls[0].kwargs["data"]
+        == cluster.endpoint.request.mock_calls[2].kwargs["data"]
     )
 
 
@@ -1055,14 +1079,14 @@ async def test_zcl_request_direction():
 
     # Input cluster
     await ep.in_clusters[zcl.clusters.general.OnOff.cluster_id].on()
-    hdr1, _ = foundation.ZCLHeader.deserialize(ep.request.mock_calls[0].args[2])
+    hdr1, _ = foundation.ZCLHeader.deserialize(ep.request.mock_calls[0].kwargs["data"])
     assert hdr1.direction == foundation.Direction.Client_to_Server
 
     ep.request.reset_mock()
 
     # Output cluster
     await ep.out_clusters[zcl.clusters.general.OnOff.cluster_id].on()
-    hdr2, _ = foundation.ZCLHeader.deserialize(ep.request.mock_calls[0].args[2])
+    hdr2, _ = foundation.ZCLHeader.deserialize(ep.request.mock_calls[0].kwargs["data"])
     assert hdr2.direction == foundation.Direction.Server_to_Client
 
     # Color cluster that also uses `direction` as a kwarg
@@ -1122,3 +1146,50 @@ async def test_zcl_reply_direction(app_mock):
     # The direction is correct
     packet_hdr, _ = foundation.ZCLHeader.deserialize(packet.data.serialize())
     assert packet_hdr.direction == foundation.Direction.Client_to_Server
+
+
+async def test_zcl_cluster_definition_backwards_compatibility():
+    class TestCluster(zcl.Cluster):
+        cluster_id = 0xABCD
+        ep_attribute = "test_cluster"
+
+        attributes = {
+            0x1234: ("attribute", t.uint8_t),
+            0x1235: ("attribute2", t.uint32_t, True),
+        }
+
+        server_commands = {
+            0x00: ("server_command", (t.uint8_t,), True),
+        }
+
+        client_commands = {
+            0x01: ("client_command", (t.uint8_t, t.uint16_t), False),
+        }
+
+    assert TestCluster.cluster_id == 0xABCD
+
+    assert TestCluster.AttributeDefs.attribute.id == 0x1234
+    assert TestCluster.AttributeDefs.attribute.type == t.uint8_t
+    assert TestCluster.AttributeDefs.attribute.is_manufacturer_specific is False
+
+    assert TestCluster.AttributeDefs.attribute2.id == 0x1235
+    assert TestCluster.AttributeDefs.attribute2.type == t.uint32_t
+    assert TestCluster.AttributeDefs.attribute2.is_manufacturer_specific is True
+
+    assert TestCluster.ServerCommandDefs.server_command.id == 0x00
+    assert len(TestCluster.ServerCommandDefs.server_command.schema.fields) == 1
+    assert (
+        TestCluster.ServerCommandDefs.server_command.schema.fields.param1.type
+        == t.uint8_t
+    )
+
+    assert TestCluster.ClientCommandDefs.client_command.id == 0x01
+    assert len(TestCluster.ClientCommandDefs.client_command.schema.fields) == 2
+    assert (
+        TestCluster.ClientCommandDefs.client_command.schema.fields.param1.type
+        == t.uint8_t
+    )
+    assert (
+        TestCluster.ClientCommandDefs.client_command.schema.fields.param2.type
+        == t.uint16_t
+    )
