@@ -3,6 +3,7 @@ from __future__ import annotations
 import abc
 import asyncio
 import collections
+from collections.abc import Coroutine
 import contextlib
 import errno
 import logging
@@ -11,7 +12,7 @@ import random
 import sys
 import time
 import typing
-from typing import Any, Coroutine, TypeVar
+from typing import Any, TypeVar
 import warnings
 
 if sys.version_info[:2] < (3, 11):
@@ -19,10 +20,11 @@ if sys.version_info[:2] < (3, 11):
 else:
     from asyncio import timeout as asyncio_timeout  # pragma: no cover
 
-from zigpy import const
 import zigpy.appdb
 import zigpy.backups
 import zigpy.config as conf
+from zigpy.const import INTERFERENCE_MESSAGE
+from zigpy.datastructures import PriorityDynamicBoundedSemaphore
 import zigpy.device
 import zigpy.endpoint
 import zigpy.exceptions
@@ -74,7 +76,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
 
         self._watchdog_task: asyncio.Task | None = None
 
-        self._concurrent_requests_semaphore = zigpy.util.DynamicBoundedSemaphore(
+        self._concurrent_requests_semaphore = PriorityDynamicBoundedSemaphore(
             self._config[conf.CONF_MAX_CONCURRENT_REQUESTS]
         )
 
@@ -186,7 +188,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
             # Some radios (like the Conbee) can fail to deliver the startup broadcast
             # due to interference
             LOGGER.warning("Failed to send startup broadcast: %s", e)
-            LOGGER.warning(const.INTERFERENCE_MESSAGE)
+            LOGGER.warning(INTERFERENCE_MESSAGE)
 
         if self.config[conf.CONF_STARTUP_ENERGY_SCAN]:
             # Each scan period is 15.36ms. Scan for at least 200ms (2^4 + 1 periods) to
@@ -202,7 +204,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
                     self.state.network_info.channel,
                     100 * results[self.state.network_info.channel] / 255,
                 )
-                LOGGER.warning(const.INTERFERENCE_MESSAGE)
+                LOGGER.warning(INTERFERENCE_MESSAGE)
 
         if self.config[conf.CONF_NWK_BACKUP_ENABLED]:
             self.backups.start_periodic_backups(
@@ -640,7 +642,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
         return False
 
     @abc.abstractmethod
-    async def connect(self):
+    async def connect(self) -> None:
         """Connect to the radio hardware and verify that it is compatible with the library.
         This method should be stateless if the connection attempt fails.
         """
@@ -746,7 +748,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
             await self.add_endpoint(endpoint)
 
     @contextlib.asynccontextmanager
-    async def _limit_concurrency(self):
+    async def _limit_concurrency(self, *, priority: int = t.PacketPriority.NORMAL):
         """Async context manager to limit global coordinator request concurrency."""
 
         start_time = time.monotonic()
@@ -759,7 +761,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
                 self._concurrent_requests_semaphore.num_waiting,
             )
 
-        async with self._concurrent_requests_semaphore:
+        async with self._concurrent_requests_semaphore(priority=priority):
             if was_locked:
                 LOGGER.debug(
                     "Previously delayed request is now running, delayed by %0.2fs",
@@ -796,6 +798,8 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
         expect_reply: bool = True,
         use_ieee: bool = False,
         extended_timeout: bool = False,
+        ask_for_ack: bool | None = None,
+        priority: int = t.PacketPriority.NORMAL,
     ) -> tuple[zigpy.zcl.foundation.Status, str]:
         """Submit and send data out as an unicast transmission.
         :param device: destination device
@@ -838,7 +842,11 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
 
         tx_options = t.TransmitOptions.NONE
 
-        if not expect_reply:
+        if ask_for_ack is not None:
+            # Prefer `ask_for_ack` to `expect_reply`
+            if ask_for_ack:
+                tx_options |= t.TransmitOptions.ACK
+        elif not expect_reply:
             tx_options |= t.TransmitOptions.ACK
 
         await self.send_packet(
@@ -854,6 +862,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
                 extended_timeout=extended_timeout,
                 source_route=source_route,
                 tx_options=tx_options,
+                priority=priority,
             )
         )
 
@@ -870,6 +879,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
         *,
         hops: int = 0,
         non_member_radius: int = 3,
+        priority: int = t.PacketPriority.NORMAL,
     ):
         """Submit and send data out as a multicast transmission.
         :param group_id: destination multicast address
@@ -899,6 +909,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
                 tx_options=t.TransmitOptions.NONE,
                 radius=hops,
                 non_member_radius=non_member_radius,
+                priority=priority,
             )
         )
 
@@ -915,6 +926,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
         sequence: t.uint8_t,
         data: bytes,
         broadcast_address: t.BroadcastAddress = t.BroadcastAddress.RX_ON_WHEN_IDLE,
+        priority: int = t.PacketPriority.NORMAL,
     ) -> tuple[zigpy.zcl.foundation.Status, str]:
         """Submit and send data out as an unicast transmission.
         :param profile: Zigbee Profile ID to use for outgoing message
@@ -945,6 +957,7 @@ class ControllerApplication(zigpy.util.ListenableMixin, abc.ABC):
                 data=t.SerializableBytes(data),
                 tx_options=t.TransmitOptions.NONE,
                 radius=radius,
+                priority=priority,
             )
         )
 
