@@ -1,11 +1,13 @@
 from __future__ import annotations
 
 import asyncio
+import contextlib
 from datetime import datetime, timezone
 import enum
 import itertools
 import logging
 import sys
+import time
 import typing
 import warnings
 
@@ -19,6 +21,8 @@ else:
 
 from zigpy import zdo
 from zigpy.const import (
+    APS_REPLY_TIMEOUT,
+    APS_REPLY_TIMEOUT_EXTENDED,
     SIG_ENDPOINTS,
     SIG_EP_INPUT,
     SIG_EP_OUTPUT,
@@ -45,9 +49,8 @@ if typing.TYPE_CHECKING:
 
 LOGGER = logging.getLogger(__name__)
 
-APS_REPLY_TIMEOUT = 5
-APS_REPLY_TIMEOUT_EXTENDED = 28
 PACKET_DEBOUNCE_WINDOW = 10
+MAX_DEVICE_CONCURRENCY = 1
 
 AFTER_OTA_ATTR_READ_DELAY = 10
 OTA_RETRY_DECORATOR = zigpy.util.retryable_request(
@@ -93,9 +96,35 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         self._send_sequence: int = 0
 
         self._packet_debouncer = zigpy.datastructures.Debouncer()
+        self._concurrent_requests_semaphore = (
+            zigpy.datastructures.PriorityDynamicBoundedSemaphore(MAX_DEVICE_CONCURRENCY)
+        )
 
         # Retained for backwards compatibility, will be removed in a future release
         self.status = Status.NEW
+
+    @contextlib.asynccontextmanager
+    async def _limit_concurrency(self, *, priority: int = 0):
+        """Async context manager to limit device request concurrency."""
+
+        start_time = time.monotonic()
+        was_locked = self._concurrent_requests_semaphore.locked()
+
+        if was_locked:
+            LOGGER.debug(
+                "Device concurrency (%s) reached, delaying device request (%s enqueued)",
+                self._concurrent_requests_semaphore.max_value,
+                self._concurrent_requests_semaphore.num_waiting,
+            )
+
+        async with self._concurrent_requests_semaphore(priority=priority):
+            if was_locked:
+                LOGGER.debug(
+                    "Previously delayed device request is now running, delayed by %0.2fs",
+                    time.monotonic() - start_time,
+                )
+
+            yield
 
     def get_sequence(self) -> t.uint8_t:
         self._send_sequence = (self._send_sequence + 1) % 256
@@ -188,7 +217,10 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
     async def get_node_descriptor(self) -> zdo_t.NodeDescriptor:
         self.info("Requesting 'Node Descriptor'")
 
-        status, _, node_desc = await self.zdo.Node_Desc_req(self.nwk)
+        status, _, node_desc = await self.zdo.Node_Desc_req(
+            self.nwk,
+            priority=t.PacketPriority.HIGH,
+        )
 
         if status != zdo_t.Status.SUCCESS:
             raise zigpy.exceptions.InvalidResponse(
@@ -231,7 +263,9 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         else:
             self.info("Discovering endpoints")
 
-            status, _, endpoints = await self.zdo.Active_EP_req(self.nwk)
+            status, _, endpoints = await self.zdo.Active_EP_req(
+                self.nwk, priority=t.PacketPriority.HIGH
+            )
 
             if status != zdo_t.Status.SUCCESS:
                 raise zigpy.exceptions.InvalidResponse(
@@ -308,6 +342,8 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         expect_reply=True,
         timeout=APS_REPLY_TIMEOUT,
         use_ieee=False,
+        ask_for_ack: bool | None = None,
+        priority: int = t.PacketPriority.NORMAL,
     ):
         extended_timeout = False
 
@@ -318,28 +354,31 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
 
         # Use a lambda so we don't leave the coroutine unawaited in case of an exception
         send_request = lambda: self._application.request(  # noqa: E731
-            self,
-            profile,
-            cluster,
-            src_ep,
-            dst_ep,
-            sequence,
-            data,
+            device=self,
+            profile=profile,
+            cluster=cluster,
+            src_ep=src_ep,
+            dst_ep=dst_ep,
+            sequence=sequence,
+            data=data,
             expect_reply=expect_reply,
             use_ieee=use_ieee,
             extended_timeout=extended_timeout,
+            ask_for_ack=ask_for_ack,
+            priority=priority,
         )
 
-        if not expect_reply:
-            await send_request()
-            return None
+        async with self._limit_concurrency(priority=priority):
+            if not expect_reply:
+                await send_request()
+                return None
 
-        # Only create a pending request if we are expecting a reply
-        with self._pending.new(sequence) as req:
-            await send_request()
+            # Only create a pending request if we are expecting a reply
+            with self._pending.new(sequence) as req:
+                await send_request()
 
-            async with asyncio_timeout(timeout):
-                return await req.result
+                async with asyncio_timeout(timeout):
+                    return await req.result
 
     def handle_message(
         self,
@@ -503,17 +542,31 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         )
 
     async def reply(
-        self, profile, cluster, src_ep, dst_ep, sequence, data, use_ieee=False
+        self,
+        profile,
+        cluster,
+        src_ep,
+        dst_ep,
+        sequence,
+        data,
+        timeout=APS_REPLY_TIMEOUT,
+        expect_reply: bool = False,
+        use_ieee: bool = False,
+        ask_for_ack: bool | None = None,
+        priority: int = t.PacketPriority.NORMAL,
     ):
         return await self.request(
-            profile,
-            cluster,
-            src_ep,
-            dst_ep,
-            sequence,
-            data,
-            expect_reply=False,
+            profile=profile,
+            cluster=cluster,
+            src_ep=src_ep,
+            dst_ep=dst_ep,
+            sequence=sequence,
+            data=data,
+            expect_reply=expect_reply,
+            timeout=timeout,
             use_ieee=use_ieee,
+            ask_for_ack=ask_for_ack,
+            priority=priority,
         )
 
     async def update_firmware(
