@@ -5,6 +5,7 @@ import enum
 import logging
 from typing import Any
 
+from zigpy.const import APS_REPLY_TIMEOUT
 import zigpy.exceptions
 import zigpy.profiles
 import zigpy.types as t
@@ -61,7 +62,7 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
             self.info("Endpoint descriptor already queried")
         else:
             status, _, sd = await self._device.zdo.Simple_Desc_req(
-                self._device.nwk, self._endpoint_id
+                self._device.nwk, self._endpoint_id, priority=t.PacketPriority.HIGH
             )
 
             if status == ZDOStatus.NOT_ACTIVE:
@@ -199,7 +200,7 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         for names in (["manufacturer", "model"], ["manufacturer"], ["model"]):
             try:
                 success, failure = await self.basic.read_attributes(
-                    names, allow_cache=True
+                    names, allow_cache=True, priority=t.PacketPriority.HIGH
                 )
             except asyncio.TimeoutError:
                 # Only swallow the `TimeoutError` on the double attribute read
@@ -251,8 +252,12 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         cluster: t.ClusterId,
         sequence: t.uint8_t,
         data: bytes,
-        expect_reply: bool = True,
         command_id: GeneralCommand | t.uint8_t = 0x00,
+        timeout=APS_REPLY_TIMEOUT,
+        expect_reply: bool = True,
+        use_ieee: bool = False,
+        ask_for_ack: bool | None = None,
+        priority: int = t.PacketPriority.NORMAL,
     ):
         if self.profile_id == zigpy.profiles.zll.PROFILE_ID and not (
             cluster == zigpy.zcl.clusters.lightlink.LightLink.cluster_id
@@ -263,13 +268,17 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
             profile_id = self.profile_id
 
         return await self.device.request(
-            profile_id,
-            cluster,
-            self._endpoint_id,
-            self._endpoint_id,
-            sequence,
-            data,
+            profile=profile_id,
+            cluster=cluster,
+            src_ep=self._endpoint_id,
+            dst_ep=self._endpoint_id,
+            sequence=sequence,
+            data=data,
+            timeout=timeout,
             expect_reply=expect_reply,
+            use_ieee=use_ieee,
+            ask_for_ack=ask_for_ack,
+            priority=priority,
         )
 
     async def reply(
@@ -278,6 +287,11 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         sequence: t.uint8_t,
         data: bytes,
         command_id: GeneralCommand | t.uint8_t = 0x00,
+        timeout=APS_REPLY_TIMEOUT,
+        expect_reply: bool = False,
+        use_ieee: bool = False,
+        ask_for_ack: bool | None = None,
+        priority: int = t.PacketPriority.NORMAL,
     ) -> None:
         if self.profile_id == zigpy.profiles.zll.PROFILE_ID and not (
             cluster == zigpy.zcl.clusters.lightlink.LightLink.cluster_id
@@ -288,7 +302,17 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
             profile_id = self.profile_id
 
         return await self.device.reply(
-            profile_id, cluster, self._endpoint_id, self._endpoint_id, sequence, data
+            profile=profile_id,
+            cluster=cluster,
+            src_ep=self._endpoint_id,
+            dst_ep=self._endpoint_id,
+            sequence=sequence,
+            data=data,
+            timeout=timeout,
+            expect_reply=expect_reply,
+            use_ieee=use_ieee,
+            ask_for_ack=ask_for_ack,
+            priority=priority,
         )
 
     def log(self, lvl: int, msg: str, *args: Any, **kwargs: Any) -> None:
