@@ -3,9 +3,12 @@
 from __future__ import annotations
 
 import collections
+from copy import deepcopy
+import dataclasses
 from enum import Enum
 import inspect
 import logging
+import pathlib
 import typing
 from typing import TYPE_CHECKING, Any
 
@@ -21,7 +24,7 @@ from zigpy.const import (
     SIG_NODE_DESC,
     SIG_SKIP_CONFIG,
 )
-from zigpy.quirks import _DEVICE_REGISTRY, CustomCluster, CustomDevice, FilterType
+from zigpy.quirks import _DEVICE_REGISTRY, BaseCustomDevice, CustomCluster, FilterType
 from zigpy.quirks.registry import DeviceRegistry
 from zigpy.quirks.v2.homeassistant import EntityPlatform, EntityType
 from zigpy.quirks.v2.homeassistant.binary_sensor import BinarySensorDeviceClass
@@ -49,7 +52,16 @@ UNBUILT_QUIRK_BUILDERS: list[QuirkBuilder] = []
 # pylint: disable=too-few-public-methods
 
 
-class CustomDeviceV2(CustomDevice):
+@dataclasses.dataclass(frozen=True)
+class ReportingConfig:
+    """Reporting config for an entity attribute."""
+
+    min_interval: int
+    max_interval: int
+    reportable_change: int
+
+
+class CustomDeviceV2(BaseCustomDevice):
     """Implementation of a quirks v2 custom device."""
 
     _copy_cluster_attr_cache = True
@@ -82,6 +94,11 @@ class CustomDeviceV2(CustomDevice):
 
         for replace_meta in quirk_metadata.replaces_metadata:
             replace_meta(self)
+
+        for (
+            replace_occurrences_meta
+        ) in quirk_metadata.replaces_cluster_occurrences_metadata:
+            replace_occurrences_meta(self)
 
         for entity_meta in quirk_metadata.entity_metadata:
             entity_meta(self)
@@ -128,26 +145,6 @@ class CustomDeviceV2(CustomDevice):
         The value is a list of EntityMetadata instances.
         """
         return self._exposes_metadata
-
-    async def apply_custom_configuration(self, *args, **kwargs):
-        """Hook for applications to instruct instances to apply custom configuration."""
-        for endpoint in self.endpoints.values():
-            if isinstance(endpoint, ZDO):
-                continue
-            for cluster in endpoint.in_clusters.values():
-                if (
-                    isinstance(cluster, CustomCluster)
-                    and cluster.apply_custom_configuration
-                    != CustomCluster.apply_custom_configuration
-                ):
-                    await cluster.apply_custom_configuration(*args, **kwargs)
-            for cluster in endpoint.out_clusters.values():
-                if (
-                    isinstance(cluster, CustomCluster)
-                    and cluster.apply_custom_configuration
-                    != CustomCluster.apply_custom_configuration
-                ):
-                    await cluster.apply_custom_configuration(*args, **kwargs)
 
 
 @attrs.define(frozen=True, kw_only=True, repr=True)
@@ -216,6 +213,36 @@ class ReplacesMetadata:
 
 
 @attrs.define(frozen=True, kw_only=True, repr=True)
+class ReplaceClusterOccurrencesMetadata:
+    """Replaces metadata for replacing all occurrences of a cluster on a device."""
+
+    cluster_types: tuple[ClusterType] = attrs.field()
+    cluster: type[Cluster | CustomCluster] = attrs.field()
+
+    def __call__(self, device: CustomDeviceV2) -> None:
+        """Process the replace."""
+        for endpoint in device.endpoints.values():
+            if isinstance(endpoint, ZDO):
+                continue
+            if (
+                ClusterType.Server in self.cluster_types
+                and self.cluster.cluster_id in endpoint.in_clusters
+            ):
+                endpoint.in_clusters.pop(self.cluster.cluster_id)
+                endpoint.add_input_cluster(
+                    self.cluster.cluster_id, self.cluster(endpoint)
+                )
+            if (
+                ClusterType.Client in self.cluster_types
+                and self.cluster.cluster_id in endpoint.out_clusters
+            ):
+                endpoint.out_clusters.pop(self.cluster.cluster_id)
+                endpoint.add_output_cluster(
+                    self.cluster.cluster_id, self.cluster(endpoint, is_server=False)
+                )
+
+
+@attrs.define(frozen=True, kw_only=True, repr=True)
 class EntityMetadata:
     """Metadata for an exposed entity."""
 
@@ -227,8 +254,10 @@ class EntityMetadata:
     initially_disabled: bool = attrs.field(default=False)
     attribute_initialized_from_cache: bool = attrs.field(default=True)
     translation_key: str | None = attrs.field(default=None)
+    fallback_name: str = attrs.field(validator=attrs.validators.instance_of(str))
 
     def __attrs_post_init__(self) -> None:
+        """Validate the entity metadata."""
         self._validate()
 
     def __call__(self, device: CustomDeviceV2) -> None:
@@ -240,17 +269,10 @@ class EntityMetadata:
 
     def _validate(self) -> None:
         """Validate the entity metadata."""
-        has_unit: bool = hasattr(self, "unit") and getattr(self, "unit") is not None
-        has_device_class: bool = hasattr(self, "device_class") and (
-            getattr(self, "device_class") is not None
-        )
-        if has_device_class and has_unit:
+        has_device_class: bool = getattr(self, "device_class", None) is not None
+        if self.translation_key is None and not has_device_class:
             raise ValueError(
-                f"EntityMetadata cannot have both unit and device_class: {self}"
-            )
-        if has_device_class and self.translation_key is not None:
-            raise ValueError(
-                f"EntityMetadata cannot have both a translation_key and a device_class: {self}"
+                f"EntityMetadata must have a translation_key or device_class: {self}"
             )
 
 
@@ -260,6 +282,7 @@ class ZCLEnumMetadata(EntityMetadata):
 
     enum: type[Enum] = attrs.field()
     attribute_name: str = attrs.field()
+    reporting_config: ReportingConfig | None = attrs.field(default=None)
 
 
 @attrs.define(frozen=True, kw_only=True, repr=True)
@@ -267,6 +290,7 @@ class ZCLSensorMetadata(EntityMetadata):
     """Metadata for exposed ZCL attribute based sensor entity."""
 
     attribute_name: str | None = attrs.field(default=None)
+    reporting_config: ReportingConfig | None = attrs.field(default=None)
     divisor: int | None = attrs.field(default=None)
     multiplier: int | None = attrs.field(default=None)
     unit: str | None = attrs.field(default=None)
@@ -279,6 +303,7 @@ class SwitchMetadata(EntityMetadata):
     """Metadata for exposed switch entity."""
 
     attribute_name: str = attrs.field()
+    reporting_config: ReportingConfig | None = attrs.field(default=None)
     force_inverted: bool = attrs.field(default=False)
     invert_attribute_name: str | None = attrs.field(default=None)
     off_value: int = attrs.field(default=0)
@@ -290,6 +315,7 @@ class NumberMetadata(EntityMetadata):
     """Metadata for exposed number entity."""
 
     attribute_name: str = attrs.field()
+    reporting_config: ReportingConfig | None = attrs.field(default=None)
     min: float | None = attrs.field(default=None)
     max: float | None = attrs.field(default=None)
     step: float | None = attrs.field(default=None)
@@ -304,6 +330,7 @@ class BinarySensorMetadata(EntityMetadata):
     """Metadata for exposed binary sensor entity."""
 
     attribute_name: str = attrs.field()
+    reporting_config: ReportingConfig | None = attrs.field(default=None)
     device_class: BinarySensorDeviceClass | None = attrs.field(default=None)
 
 
@@ -333,13 +360,40 @@ class ManufacturerModelMetadata:
 
 
 @attrs.define(frozen=True, kw_only=True, repr=True)
+class FriendlyNameMetadata:
+    """Metadata to rename a device."""
+
+    model: str = attrs.field()
+    manufacturer: str = attrs.field()
+
+
+class DeviceAlertLevel(Enum):
+    """Device alert level."""
+
+    INFO = "info"
+    WARNING = "warning"
+    ERROR = "error"
+
+
+@attrs.define(frozen=True, kw_only=True, repr=True)
+class DeviceAlertMetadata:
+    """Metadata for device-specific alerts."""
+
+    level: DeviceAlertLevel = attrs.field(converter=DeviceAlertLevel)
+    message: str = attrs.field()
+
+
+@attrs.define(frozen=True, kw_only=True, repr=True)
 class QuirksV2RegistryEntry:
     """Quirks V2 registry entry."""
 
-    quirk_location: str = attrs.field(default=None, eq=False)
+    quirk_file: str = attrs.field(default=None, eq=False)
+    quirk_file_line: int = attrs.field(default=None, eq=False)
     manufacturer_model_metadata: tuple[ManufacturerModelMetadata] = attrs.field(
         factory=tuple
     )
+    friendly_name: FriendlyNameMetadata | None = attrs.field(default=None)
+    device_alerts: tuple[DeviceAlertMetadata] = attrs.field(factory=tuple)
     filters: tuple[FilterType] = attrs.field(factory=tuple)
     custom_device_class: type[CustomDeviceV2] | None = attrs.field(default=None)
     device_node_descriptor: NodeDescriptor | None = attrs.field(default=None)
@@ -347,6 +401,9 @@ class QuirksV2RegistryEntry:
     adds_metadata: tuple[AddsMetadata] = attrs.field(factory=tuple)
     removes_metadata: tuple[RemovesMetadata] = attrs.field(factory=tuple)
     replaces_metadata: tuple[ReplacesMetadata] = attrs.field(factory=tuple)
+    replaces_cluster_occurrences_metadata: tuple[ReplaceClusterOccurrencesMetadata] = (
+        attrs.field(factory=tuple)
+    )
     entity_metadata: tuple[
         ZCLEnumMetadata
         | SwitchMetadata
@@ -376,11 +433,21 @@ class QuirkBuilder:
     """Quirks V2 registry entry."""
 
     def __init__(
-        self, manufacturer: str, model: str, registry: DeviceRegistry = _DEVICE_REGISTRY
+        self,
+        manufacturer: str | None = None,
+        model: str | None = None,
+        registry: DeviceRegistry = _DEVICE_REGISTRY,
     ) -> None:
         """Initialize the quirk builder."""
+        if manufacturer and not model or model and not manufacturer:
+            raise ValueError(
+                "manufacturer and model must be provided together or completely omitted."
+            )
+
         self.registry: DeviceRegistry = registry
         self.manufacturer_model_metadata: list[ManufacturerModelMetadata] = []
+        self.friendly_name_metadata: FriendlyNameMetadata | None = None
+        self.device_alerts: list[DeviceAlertMetadata] = []
         self.filters: list[FilterType] = []
         self.custom_device_class: type[CustomDeviceV2] | None = None
         self.device_node_descriptor: NodeDescriptor | None = None
@@ -388,6 +455,9 @@ class QuirkBuilder:
         self.adds_metadata: list[AddsMetadata] = []
         self.removes_metadata: list[RemovesMetadata] = []
         self.replaces_metadata: list[ReplacesMetadata] = []
+        self.replaces_cluster_occurrences_metadata: list[
+            ReplaceClusterOccurrencesMetadata
+        ] = []
         self.entity_metadata: list[
             ZCLEnumMetadata
             | SwitchMetadata
@@ -402,21 +472,23 @@ class QuirkBuilder:
 
         stack: list[inspect.FrameInfo] = inspect.stack()
         caller: inspect.FrameInfo = stack[1]
-        self.quirk_location: str | None = (
-            f"file[{caller.filename}]-line:{caller.lineno}"
-        )
+        self.quirk_file = pathlib.Path(caller.filename)
+        self.quirk_file_line = caller.lineno
 
-        self.also_applies_to(manufacturer, model)
+        if manufacturer and model:
+            self.applies_to(manufacturer, model)
+
         UNBUILT_QUIRK_BUILDERS.append(self)
 
-    def also_applies_to(self, manufacturer: str, model: str) -> QuirkBuilder:
-        """Register this quirks v2 entry for an additional manufacturer and model."""
+    def applies_to(self, manufacturer: str, model: str) -> QuirkBuilder:
+        """Register this quirks v2 entry for the specified manufacturer and model."""
         self.manufacturer_model_metadata.append(
-            ManufacturerModelMetadata(  # type: ignore[call-arg]
-                manufacturer=manufacturer, model=model
-            )
+            ManufacturerModelMetadata(manufacturer=manufacturer, model=model)
         )
         return self
+
+    # backward compatibility
+    also_applies_to = applies_to
 
     def filter(self, filter_function: FilterType) -> QuirkBuilder:
         """Add a filter and returns self.
@@ -478,7 +550,7 @@ class QuirkBuilder:
         instances and their values. These attributes will be added to the cluster when
         the quirk is applied and the values will be constant.
         """
-        add = AddsMetadata(  # type: ignore[call-arg]
+        add = AddsMetadata(
             endpoint_id=endpoint_id,
             cluster=cluster,
             cluster_type=cluster_type,
@@ -497,7 +569,7 @@ class QuirkBuilder:
 
         This method allows removing a cluster from a device when the quirk is applied.
         """
-        remove = RemovesMetadata(  # type: ignore[call-arg]
+        remove = RemovesMetadata(
             endpoint_id=endpoint_id,
             cluster_id=cluster_id,
             cluster_type=cluster_type,
@@ -523,20 +595,52 @@ class QuirkBuilder:
         be removed. If cluster_id is not provided, the cluster_id of the replacement
         cluster will be used.
         """
-        remove = RemovesMetadata(  # type: ignore[call-arg]
+        remove = RemovesMetadata(
             endpoint_id=endpoint_id,
             cluster_id=cluster_id
             if cluster_id is not None
             else replacement_cluster_class.cluster_id,
             cluster_type=cluster_type,
         )
-        add = AddsMetadata(  # type: ignore[call-arg]
+        add = AddsMetadata(
             endpoint_id=endpoint_id,
             cluster=replacement_cluster_class,
             cluster_type=cluster_type,
         )
-        replace = ReplacesMetadata(remove=remove, add=add)  # type: ignore[call-arg]
+        replace = ReplacesMetadata(remove=remove, add=add)
         self.replaces_metadata.append(replace)
+        return self
+
+    def replace_cluster_occurrences(
+        self,
+        replacement_cluster_class: type[Cluster | CustomCluster],
+        replace_server_instances: bool = True,
+        replace_client_instances: bool = True,
+    ) -> QuirkBuilder:
+        """Add a ReplaceClusterOccurrencesMetadata entry and returns self.
+
+        This method allows replacing a cluster on a device across all endpoints
+        for the specified cluster types when the quirk is applied.
+
+        replacement_cluster_class should be a subclass of Cluster or CustomCluster and
+        will be used to create a new cluster instance to replace the existing cluster.
+
+        replace_server_instances and replace_client_instances control the cluster types
+        that will be replaced. If replace_server_instances is True, all server instances
+        of the cluster will be replaced. If replace_client_instances is True, all client
+        instances of the cluster will be replaced.
+        """
+        types = []
+        if replace_server_instances:
+            types.append(ClusterType.Server)
+        if replace_client_instances:
+            types.append(ClusterType.Client)
+        self.replaces_cluster_occurrences_metadata.append(
+            ReplaceClusterOccurrencesMetadata(
+                cluster_types=tuple(types),
+                cluster=replacement_cluster_class,
+            )
+        )
         return self
 
     def enum(
@@ -546,26 +650,30 @@ class QuirkBuilder:
         cluster_id: int,
         cluster_type: ClusterType = ClusterType.Server,
         endpoint_id: int = 1,
-        entity_type: EntityType = EntityType.CONFIG,
         entity_platform: EntityPlatform = EntityPlatform.SELECT,
+        entity_type: EntityType = EntityType.CONFIG,
         initially_disabled: bool = False,
         attribute_initialized_from_cache: bool = True,
+        reporting_config: ReportingConfig | None = None,
         translation_key: str | None = None,
+        fallback_name: str | None = None,
     ) -> QuirkBuilder:
         """Add an EntityMetadata containing ZCLEnumMetadata and return self.
 
         This method allows exposing an enum based entity in Home Assistant.
         """
         self.entity_metadata.append(
-            ZCLEnumMetadata(  # type: ignore[call-arg]
+            ZCLEnumMetadata(
                 endpoint_id=endpoint_id,
                 cluster_id=cluster_id,
                 cluster_type=cluster_type,
-                entity_type=entity_type,
                 entity_platform=entity_platform,
+                entity_type=entity_type,
                 initially_disabled=initially_disabled,
                 attribute_initialized_from_cache=attribute_initialized_from_cache,
+                reporting_config=reporting_config,
                 translation_key=translation_key,
+                fallback_name=fallback_name,
                 enum=enum_class,
                 attribute_name=attribute_name,
             )
@@ -586,14 +694,16 @@ class QuirkBuilder:
         unit: str | None = None,
         initially_disabled: bool = False,
         attribute_initialized_from_cache: bool = True,
+        reporting_config: ReportingConfig | None = None,
         translation_key: str | None = None,
+        fallback_name: str | None = None,
     ) -> QuirkBuilder:
         """Add an EntityMetadata containing ZCLSensorMetadata and return self.
 
         This method allows exposing a sensor entity in Home Assistant.
         """
         self.entity_metadata.append(
-            ZCLSensorMetadata(  # type: ignore[call-arg]
+            ZCLSensorMetadata(
                 endpoint_id=endpoint_id,
                 cluster_id=cluster_id,
                 cluster_type=cluster_type,
@@ -601,7 +711,9 @@ class QuirkBuilder:
                 entity_type=entity_type,
                 initially_disabled=initially_disabled,
                 attribute_initialized_from_cache=attribute_initialized_from_cache,
+                reporting_config=reporting_config,
                 translation_key=translation_key,
+                fallback_name=fallback_name,
                 attribute_name=attribute_name,
                 divisor=divisor,
                 multiplier=multiplier,
@@ -623,24 +735,29 @@ class QuirkBuilder:
         off_value: int = 0,
         on_value: int = 1,
         entity_platform=EntityPlatform.SWITCH,
+        entity_type: EntityType = EntityType.CONFIG,
         initially_disabled: bool = False,
         attribute_initialized_from_cache: bool = True,
+        reporting_config: ReportingConfig | None = None,
         translation_key: str | None = None,
+        fallback_name: str | None = None,
     ) -> QuirkBuilder:
         """Add an EntityMetadata containing SwitchMetadata and return self.
 
         This method allows exposing a switch entity in Home Assistant.
         """
         self.entity_metadata.append(
-            SwitchMetadata(  # type: ignore[call-arg]
+            SwitchMetadata(
                 endpoint_id=endpoint_id,
                 cluster_id=cluster_id,
                 cluster_type=cluster_type,
                 entity_platform=entity_platform,
-                entity_type=EntityType.CONFIG,
+                entity_type=entity_type,
                 initially_disabled=initially_disabled,
                 attribute_initialized_from_cache=attribute_initialized_from_cache,
+                reporting_config=reporting_config,
                 translation_key=translation_key,
+                fallback_name=fallback_name,
                 attribute_name=attribute_name,
                 force_inverted=force_inverted,
                 invert_attribute_name=invert_attribute_name,
@@ -662,25 +779,30 @@ class QuirkBuilder:
         unit: str | None = None,
         mode: str | None = None,
         multiplier: float | None = None,
+        entity_type: EntityType = EntityType.CONFIG,
         device_class: NumberDeviceClass | None = None,
         initially_disabled: bool = False,
         attribute_initialized_from_cache: bool = True,
+        reporting_config: ReportingConfig | None = None,
         translation_key: str | None = None,
+        fallback_name: str | None = None,
     ) -> QuirkBuilder:
         """Add an EntityMetadata containing NumberMetadata and return self.
 
         This method allows exposing a number entity in Home Assistant.
         """
         self.entity_metadata.append(
-            NumberMetadata(  # type: ignore[call-arg]
+            NumberMetadata(
                 endpoint_id=endpoint_id,
                 cluster_id=cluster_id,
                 cluster_type=cluster_type,
                 entity_platform=EntityPlatform.NUMBER,
-                entity_type=EntityType.CONFIG,
+                entity_type=entity_type,
                 initially_disabled=initially_disabled,
                 attribute_initialized_from_cache=attribute_initialized_from_cache,
+                reporting_config=reporting_config,
                 translation_key=translation_key,
+                fallback_name=fallback_name,
                 attribute_name=attribute_name,
                 min=min_value,
                 max=max_value,
@@ -699,25 +821,30 @@ class QuirkBuilder:
         cluster_id: int,
         cluster_type: ClusterType = ClusterType.Server,
         endpoint_id: int = 1,
+        entity_type: EntityType = EntityType.DIAGNOSTIC,
         device_class: BinarySensorDeviceClass | None = None,
         initially_disabled: bool = False,
         attribute_initialized_from_cache: bool = True,
+        reporting_config: ReportingConfig | None = None,
         translation_key: str | None = None,
+        fallback_name: str | None = None,
     ) -> QuirkBuilder:
         """Add an EntityMetadata containing BinarySensorMetadata and return self.
 
         This method allows exposing a binary sensor entity in Home Assistant.
         """
         self.entity_metadata.append(
-            BinarySensorMetadata(  # type: ignore[call-arg]
+            BinarySensorMetadata(
                 endpoint_id=endpoint_id,
                 cluster_id=cluster_id,
                 cluster_type=cluster_type,
                 entity_platform=EntityPlatform.BINARY_SENSOR,
-                entity_type=EntityType.DIAGNOSTIC,
+                entity_type=entity_type,
                 initially_disabled=initially_disabled,
                 attribute_initialized_from_cache=attribute_initialized_from_cache,
+                reporting_config=reporting_config,
                 translation_key=translation_key,
+                fallback_name=fallback_name,
                 attribute_name=attribute_name,
                 device_class=device_class,
             )
@@ -735,6 +862,7 @@ class QuirkBuilder:
         initially_disabled: bool = False,
         attribute_initialized_from_cache: bool = True,
         translation_key: str | None = None,
+        fallback_name: str | None = None,
     ) -> QuirkBuilder:
         """Add an EntityMetadata containing WriteAttributeButtonMetadata and return self.
 
@@ -742,7 +870,7 @@ class QuirkBuilder:
         a value to an attribute when pressed.
         """
         self.entity_metadata.append(
-            WriteAttributeButtonMetadata(  # type: ignore[call-arg]
+            WriteAttributeButtonMetadata(
                 endpoint_id=endpoint_id,
                 cluster_id=cluster_id,
                 cluster_type=cluster_type,
@@ -751,6 +879,7 @@ class QuirkBuilder:
                 initially_disabled=initially_disabled,
                 attribute_initialized_from_cache=attribute_initialized_from_cache,
                 translation_key=translation_key,
+                fallback_name=fallback_name,
                 attribute_name=attribute_name,
                 attribute_value=attribute_value,
             )
@@ -768,6 +897,7 @@ class QuirkBuilder:
         entity_type: EntityType = EntityType.CONFIG,
         initially_disabled: bool = False,
         translation_key: str | None = None,
+        fallback_name: str | None = None,
     ) -> QuirkBuilder:
         """Add an EntityMetadata containing ZCLCommandButtonMetadata and return self.
 
@@ -775,7 +905,7 @@ class QuirkBuilder:
         a ZCL command when pressed.
         """
         self.entity_metadata.append(
-            ZCLCommandButtonMetadata(  # type: ignore[call-arg]
+            ZCLCommandButtonMetadata(
                 endpoint_id=endpoint_id,
                 cluster_id=cluster_id,
                 cluster_type=cluster_type,
@@ -783,6 +913,7 @@ class QuirkBuilder:
                 entity_type=entity_type,
                 initially_disabled=initially_disabled,
                 translation_key=translation_key,
+                fallback_name=fallback_name,
                 command_name=command_name,
                 args=command_args if command_args is not None else (),
                 kwargs=command_kwargs if command_kwargs is not None else frozendict(),
@@ -797,11 +928,30 @@ class QuirkBuilder:
         self.device_automation_triggers_metadata.update(device_automation_triggers)
         return self
 
+    def friendly_name(self, *, model: str, manufacturer: str) -> QuirkBuilder:
+        """Renames the device."""
+        self.friendly_name_metadata = FriendlyNameMetadata(
+            model=model, manufacturer=manufacturer
+        )
+        return self
+
+    def device_alert(self, *, level: DeviceAlertLevel, message: str) -> QuirkBuilder:
+        """Adds a device alert."""
+        self.device_alerts.append(DeviceAlertMetadata(level=level, message=message))
+        return self
+
     def add_to_registry(self) -> QuirksV2RegistryEntry:
         """Build the quirks v2 registry entry."""
-        quirk: QuirksV2RegistryEntry = QuirksV2RegistryEntry(  # type: ignore[call-arg]
+        if not self.manufacturer_model_metadata:
+            raise ValueError(
+                "At least one manufacturer and model must be specified for a v2 quirk."
+            )
+        quirk: QuirksV2RegistryEntry = QuirksV2RegistryEntry(
             manufacturer_model_metadata=tuple(self.manufacturer_model_metadata),
-            quirk_location=self.quirk_location,
+            friendly_name=self.friendly_name_metadata,
+            device_alerts=tuple(self.device_alerts),
+            quirk_file=self.quirk_file,
+            quirk_file_line=self.quirk_file_line,
             filters=tuple(self.filters),
             custom_device_class=self.custom_device_class,
             device_node_descriptor=self.device_node_descriptor,
@@ -809,6 +959,9 @@ class QuirkBuilder:
             adds_metadata=tuple(self.adds_metadata),
             removes_metadata=tuple(self.removes_metadata),
             replaces_metadata=tuple(self.replaces_metadata),
+            replaces_cluster_occurrences_metadata=tuple(
+                self.replaces_cluster_occurrences_metadata
+            ),
             entity_metadata=tuple(self.entity_metadata),
             device_automation_triggers_metadata=self.device_automation_triggers_metadata,
         )
@@ -821,6 +974,14 @@ class QuirkBuilder:
             UNBUILT_QUIRK_BUILDERS.remove(self)
 
         return quirk
+
+    def clone(self, omit_man_model_data=True) -> QuirkBuilder:
+        """Clone this QuirkBuilder potentially omitting manufacturer and model data."""
+        new_builder = deepcopy(self)
+        new_builder.registry = self.registry
+        if omit_man_model_data:
+            new_builder.manufacturer_model_metadata = []
+        return new_builder
 
 
 def add_to_registry_v2(
