@@ -1,9 +1,10 @@
 from __future__ import annotations
 
+from collections.abc import Coroutine
 import functools
 import logging
-from typing import Coroutine
 
+from zigpy.const import APS_REPLY_TIMEOUT
 import zigpy.profiles
 import zigpy.types as t
 from zigpy.typing import AddressingMode
@@ -30,8 +31,10 @@ class ZDO(zigpy.util.CatchingTaskMixin, zigpy.util.ListenableMixin):
         self._device = device
         self._listeners = {}
 
-    def _serialize(self, command, *args):
-        schema = types.CLUSTERS[command][1]
+    def _serialize(self, command, *args, **kwargs):
+        keys, schema = types.CLUSTERS[command]
+        # TODO: expose this in a future PR
+        assert not kwargs
         return t.serialize(args, schema)
 
     def deserialize(self, cluster_id, data):
@@ -48,18 +51,61 @@ class ZDO(zigpy.util.CatchingTaskMixin, zigpy.util.ListenableMixin):
 
         return hdr, args
 
-    def request(self, command, *args, use_ieee=False):
-        data = self._serialize(command, *args)
+    def request(
+        self,
+        command,
+        *args,
+        timeout=APS_REPLY_TIMEOUT,
+        expect_reply: bool = True,
+        use_ieee: bool = False,
+        ask_for_ack: bool | None = None,
+        priority: int = t.PacketPriority.NORMAL,
+        **kwargs,
+    ):
+        data = self._serialize(command, *args, **kwargs)
         tsn = self.device.get_sequence()
-        data = t.uint8_t(tsn).serialize() + data
-        return self._device.request(0, command, 0, 0, tsn, data, use_ieee=use_ieee)
+        return self._device.request(
+            profile=0x0000,
+            cluster=command,
+            src_ep=ZDO_ENDPOINT,
+            dst_ep=ZDO_ENDPOINT,
+            sequence=tsn,
+            data=t.uint8_t(tsn).serialize() + data,
+            timeout=timeout,
+            expect_reply=expect_reply,
+            use_ieee=use_ieee,
+            ask_for_ack=ask_for_ack,
+            priority=priority,
+        )
 
-    def reply(self, command, *args, tsn=None, use_ieee=False):
-        data = self._serialize(command, *args)
+    def reply(
+        self,
+        command,
+        *args,
+        tsn: int | t.uint8_t | None = None,
+        timeout=APS_REPLY_TIMEOUT,
+        expect_reply: bool = False,
+        use_ieee: bool = False,
+        ask_for_ack: bool | None = None,
+        priority: int = t.PacketPriority.NORMAL,
+        **kwargs,
+    ):
+        data = self._serialize(command, *args, **kwargs)
         if tsn is None:
             tsn = self.device.get_sequence()
-        data = t.uint8_t(tsn).serialize() + data
-        return self._device.reply(0, command, 0, 0, tsn, data, use_ieee=use_ieee)
+        return self._device.reply(
+            profile=0x0000,
+            cluster=command,
+            src_ep=ZDO_ENDPOINT,
+            dst_ep=ZDO_ENDPOINT,
+            sequence=tsn,
+            data=t.uint8_t(tsn).serialize() + data,
+            timeout=timeout,
+            expect_reply=expect_reply,
+            use_ieee=use_ieee,
+            ask_for_ack=ask_for_ack,
+            priority=priority,
+        )
 
     def handle_message(
         self,
@@ -107,6 +153,7 @@ class ZDO(zigpy.util.CatchingTaskMixin, zigpy.util.ListenableMixin):
                     0,
                     [],
                     tsn=hdr.tsn,
+                    priority=t.PacketPriority.LOW,
                 )
             )
 
@@ -136,6 +183,7 @@ class ZDO(zigpy.util.CatchingTaskMixin, zigpy.util.ListenableMixin):
                     0,
                     [],
                     tsn=hdr.tsn,
+                    priority=t.PacketPriority.LOW,
                 )
             )
 
@@ -175,12 +223,24 @@ class ZDO(zigpy.util.CatchingTaskMixin, zigpy.util.ListenableMixin):
         local_addr = self._device.application.state.node_info.nwk
         if profile != zigpy.profiles.zha.PROFILE_ID:
             self.create_catching_task(
-                self.Match_Desc_rsp(0, local_addr, [], tsn=hdr.tsn)
+                self.Match_Desc_rsp(
+                    0,
+                    local_addr,
+                    [],
+                    tsn=hdr.tsn,
+                    priority=t.PacketPriority.HIGH,
+                )
             )
             return
 
         self.create_catching_task(
-            self.Match_Desc_rsp(0, local_addr, [t.uint8_t(1)], tsn=hdr.tsn)
+            self.Match_Desc_rsp(
+                0,
+                local_addr,
+                [t.uint8_t(1)],
+                tsn=hdr.tsn,
+                priority=t.PacketPriority.HIGH,
+            )
         )
 
     def bind(self, cluster):
